requirements.txt
Flask
pandas
openai
python-dotenv
faiss-cpu
scikit-learn
sentence-transformers 